+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/geo/GEOMiner.java
@@ -17,19 +17,25 @@ import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.cscorelib2.item.CustomItem;
 import io.github.thebusybiscuit.slimefun4.api.geo.GEOResource;
+import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetComponent;
 import io.github.thebusybiscuit.slimefun4.core.attributes.HologramOwner;
+import io.github.thebusybiscuit.slimefun4.core.attributes.MachineProcessHolder;
 import io.github.thebusybiscuit.slimefun4.core.attributes.RecipeDisplayItem;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockBreakHandler;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockPlaceHandler;
+import io.github.thebusybiscuit.slimefun4.core.machines.MachineProcessor;
+import io.github.thebusybiscuit.slimefun4.core.networks.energy.EnergyNetComponentType;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.handlers.SimpleBlockBreakHandler;
-import io.github.thebusybiscuit.slimefun4.utils.ChestMenuUtils;
+import io.github.thebusybiscuit.slimefun4.implementation.operations.MiningOperation;
+import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.CSCoreLibPlugin.general.Inventory.ChestMenu.AdvancedMenuClickHandler;
 import me.mrCookieSlime.CSCoreLibPlugin.general.Inventory.ClickAction;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AContainer;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.MachineRecipe;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.interfaces.InventoryBlock;
+import me.mrCookieSlime.Slimefun.Objects.handlers.BlockTicker;
 import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
@@ -42,22 +48,33 @@ import me.mrCookieSlime.Slimefun.api.inventory.BlockMenuPreset;
  *
  * @see GEOResource
  */
-public class GEOMiner extends AContainer implements RecipeDisplayItem, HologramOwner {
+public class GEOMiner extends SlimefunItem implements RecipeDisplayItem, EnergyNetComponent, InventoryBlock, HologramOwner, MachineProcessHolder<MiningOperation> {
 
     private static final int[] BORDER = { 0, 1, 2, 3, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 26, 27, 35, 36, 44, 45, 53 };
     private static final int[] OUTPUT_BORDER = { 19, 20, 21, 22, 23, 24, 25, 28, 34, 37, 43, 46, 47, 48, 49, 50, 51, 52 };
     private static final int[] OUTPUT_SLOTS = { 29, 30, 31, 32, 33, 38, 39, 40, 41, 42 };
+
     private static final int PROCESSING_TIME = 14;
+    private static final int ENERGY_CONSUMPTION = 24;
+
+    private final MachineProcessor<MiningOperation> processor = new MachineProcessor<>(this);
 
     @ParametersAreNonnullByDefault
     public GEOMiner(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
 
-        addItemHandler(onPlace());
+        processor.setProgressBar(new ItemStack(Material.DIAMOND_PICKAXE));
+        createPreset(this, getItemName(), this::constructMenu);
+        addItemHandler(onBlockPlace(), onBlockBreak());
+    }
+
+    @Override
+    public MachineProcessor<MiningOperation> getMachineProcessor() {
+        return processor;
     }
 
     @Nonnull
-    private BlockPlaceHandler onPlace() {
+    private BlockPlaceHandler onBlockPlace() {
         return new BlockPlaceHandler(false) {
 
             @Override
@@ -68,36 +85,21 @@ public class GEOMiner extends AContainer implements RecipeDisplayItem, HologramO
     }
 
     @Nonnull
-    @Override
-    protected BlockBreakHandler onBlockBreak() {
+    private BlockBreakHandler onBlockBreak() {
         return new SimpleBlockBreakHandler() {
 
             @Override
             public void onBlockBreak(@Nonnull Block b) {
                 removeHologram(b);
-
                 BlockMenu inv = BlockStorage.getInventory(b);
 
                 if (inv != null) {
-                    inv.dropItems(b.getLocation(), getOutputSlots());
+                    inv.dropItems(b.getLocation(), OUTPUT_SLOTS);
                 }
 
-                progress.remove(b);
-                processing.remove(b);
+                processor.endOperation(b);
             }
         };
-    }
-
-    @Nonnull
-    @Override
-    public String getMachineIdentifier() {
-        return "GEO_MINER";
-    }
-
-    @Nonnull
-    @Override
-    public ItemStack getProgressBar() {
-        return new ItemStack(Material.DIAMOND_PICKAXE);
     }
 
     @Nonnull
@@ -133,6 +135,15 @@ public class GEOMiner extends AContainer implements RecipeDisplayItem, HologramO
     }
 
     @Override
+    public EnergyNetComponentType getEnergyComponentType() {
+        return EnergyNetComponentType.CONSUMER;
+    }
+
+    @Override
+    public int getCapacity() {
+        return 512;
+    }
+
     protected void constructMenu(@Nonnull BlockMenuPreset preset) {
         for (int i : BORDER) {
             preset.addItem(i, new CustomItem(Material.GRAY_STAINED_GLASS_PANE, " "), (p, slot, item, action) -> false);
@@ -144,7 +155,7 @@ public class GEOMiner extends AContainer implements RecipeDisplayItem, HologramO
 
         preset.addItem(4, new CustomItem(Material.BLACK_STAINED_GLASS_PANE, " "), (p, slot, item, action) -> false);
 
-        for (int i : getOutputSlots()) {
+        for (int i : OUTPUT_SLOTS) {
             preset.addMenuClickHandler(i, new AdvancedMenuClickHandler() {
 
                 @Override
@@ -161,27 +172,40 @@ public class GEOMiner extends AContainer implements RecipeDisplayItem, HologramO
     }
 
     @Override
+    public void preRegister() {
+        addItemHandler(new BlockTicker() {
+
+            @Override
+            public void tick(Block b, SlimefunItem sf, Config data) {
+                GEOMiner.this.tick(b);
+            }
+
+            @Override
+            public boolean isSynchronized() {
+                return false;
+            }
+        });
+    }
+
     protected void tick(@Nonnull Block b) {
         BlockMenu inv = BlockStorage.getInventory(b);
+        MiningOperation operation = processor.getOperation(b);
 
-        if (isProcessing(b)) {
-            int timeleft = progress.get(b);
+        if (operation != null) {
+            if (!operation.isFinished()) {
+                processor.updateProgressBar(inv, 4, operation);
 
-            if (timeleft > 0) {
-                ChestMenuUtils.updateProgressbar(inv, 4, timeleft, processing.get(b).getTicks(), getProgressBar());
-
-                if (getCharge(b.getLocation()) < getEnergyConsumption()) {
+                if (getCharge(b.getLocation()) < ENERGY_CONSUMPTION) {
                     return;
                 }
 
-                removeCharge(b.getLocation(), getEnergyConsumption());
-                progress.put(b, timeleft - 1);
+                removeCharge(b.getLocation(), ENERGY_CONSUMPTION);
+                operation.addProgress(1);
             } else {
                 inv.replaceExistingItem(4, new CustomItem(Material.BLACK_STAINED_GLASS_PANE, " "));
-                inv.pushItem(processing.get(b).getOutput()[0], getOutputSlots());
+                inv.pushItem(operation.getResult(), OUTPUT_SLOTS);
 
-                progress.remove(b);
-                processing.remove(b);
+                processor.endOperation(b);
             }
         } else if (!BlockStorage.hasChunkInfo(b.getWorld(), b.getX() >> 4, b.getZ() >> 4)) {
             updateHologram(b, "&4GEO-Scan required!");
@@ -202,14 +226,11 @@ public class GEOMiner extends AContainer implements RecipeDisplayItem, HologramO
 
                 int supplies = optional.getAsInt();
                 if (supplies > 0) {
-                    MachineRecipe r = new MachineRecipe(PROCESSING_TIME / getSpeed(), new ItemStack[0], new ItemStack[] { resource.getItem().clone() });
-
-                    if (!inv.fits(r.getOutput()[0], getOutputSlots())) {
+                    if (!inv.fits(resource.getItem(), OUTPUT_SLOTS)) {
                         return;
                     }
 
-                    processing.put(b, r);
-                    progress.put(b, r.getTicks());
+                    processor.startOperation(b, new MiningOperation(resource.getItem().clone(), PROCESSING_TIME));
                     SlimefunPlugin.getGPSNetwork().getResourceManager().setSupplies(resource, b.getWorld(), b.getX() >> 4, b.getZ() >> 4, supplies - 1);
                     updateHologram(b, "&7Mining: &r" + resource.getName());
                     return;